armv7a: work around Error: co-processor offset out of range
The offset of immediate FP values on various armv7a cores are limited
to PC +/-1kB. These limits are sometimes breached when using constants
in long functions and more often by the usage of assembler macros where
the constant labels are defined far from the macros' call locations.

Instead of using fixed labels for constants that refer to locations
before the respective macros this patch uses macro parameters to pass
the label name into them. For each macro that calls such a macro this
patch creates a new constant label before the caller (thus ensuring
it's near enough when actually assembling).

When the routines are defined as assembler functions (for lower
MACRO_LEVEL) this patch introduces new constants (that are defined
before the function but only if functions are used instead of macros).
They are used (instead of the macro parameters) according to the
value of MACRO_LEVEL via #if directives.

For some longer routines one even has to refer to a constant defined
after them in which case "_post" is appended to the constant's name.